# app/docs/openapi.yaml
openapi: 3.1.0
info:
  title: whisperX REST service
  description: "\n    # whisperX RESTful API\n\n    Welcome to the whisperX RESTful\
    \ API! This API provides a suite of audio processing services to enhance and analyze\
    \ your audio content.\n\n    ## Documentation:\n\n    For detailed information\
    \ on request and response formats, consult the [WhisperX Documentation](https://github.com/m-bain/whisperX).\n\
    \n    ## Services:\n\n    Speech-2-Text provides a suite of audio processing services\
    \ to enhance and analyze your audio content. The following services are available:\n\
    \n    1. Transcribe: Transcribe an audio/video  file into text.\n    2. Align:\
    \ Align the transcript to the audio/video file.\n    3. Diarize: Diarize an audio/video\
    \ file into speakers.\n    4. Combine Transcript and Diarization: Combine the\
    \ transcript and diarization results.\n\n    ## Supported file extensions:\n \
    \   AUDIO_EXTENSIONS = {'.aac', '.mp3', '.oga', '.m4a', '.wav', '.ogg', '.wma',\
    \ '.amr', '.awb', '.flac'}\n\n    VIDEO_EXTENSIONS = {'.avi', '.mkv', '.webm',\
    \ '.wmv', '.mp4', '.mov'}\n\n    ## Language-Specific Model Overrides:\n\n   \
    \ Some languages have fine-tuned models that are automatically used when that\
    \ language is selected.\n    The server-side `LANGUAGE_MODEL_OVERRIDES` configuration\
    \ maps language codes to optimized CTranslate2\n    model paths. When an override\
...
  version: 0.0.1
paths:
...
  /task/all:
    get:
      tags:
      - Tasks Management
      summary: Get All Tasks Status
      description: "Retrieve a paginated, optionally filtered list of tasks (Plan\
        \ 15-ux).\n\nPagination + search is server-side so the queue can scale beyond\
        \ a few\nhundred rows without dragging the whole table to the browser. Sync\n\
        ``def`` so FastAPI runs the query in the threadpool: a slow query never\n\
        stalls the event loop (and with it every other route). The\nuser-scope filter\
        \ (Phase 13-07) still applies — callers see only their\nown tasks.\n\nArgs:\n\
        \    q: Case-insensitive substring match against file_name.\n    status: Exact-match\
        \ status filter (processing|completed|failed).\n    page: 1-indexed page number;\
        \ Pydantic ``ge=1`` rejects 0/negatives.\n    page_size: Items per page, 1..200;\
        \ out-of-range -> 422.\n    service: Task management service dependency (scoped\
        \ to caller).\n\nReturns:\n    TaskListResponse: ``{tasks, total, page, page_size}``."
      operationId: get_all_tasks_status_task_all_get
      parameters:
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 200
          - type: 'null'
          description: Substring match against file_name (case-insensitive)
          title: Q
        description: Substring match against file_name (case-insensitive)
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 32
          - type: 'null'
          description: Filter by task status (processing|completed|failed)
          title: Status
        description: Filter by task status (processing|completed|failed)
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          description: Page number (1-indexed)
          default: 1
          title: Page
        description: Page number (1-indexed)
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          description: Items per page (1..200)
          default: 50
          title: Page Size
        description: Items per page (1..200)
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /task/{identifier}:
    get:
      tags:
      - Tasks Management
      summary: Get Transcription Status
      description: "Retrieve the status of a specific task by its identifier.\n\n\
        Args:\n    identifier (str): The identifier of the task.\n    service: Task\
        \ management service dependency.\n\nReturns:\n    Result: The status of the\
        \ task.\n\nRaises:\n    TaskNotFoundError: If the identifier is not found."
      operationId: get_transcription_status_task__identifier__get
      parameters:
      - name: identifier
        in: path
        required: true
        schema:
          type: string
          title: Identifier
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /task/{identifier}/delete:
    delete:
      tags:
      - Tasks Management
      summary: Delete Task
      description: "Delete a specific task by its identifier.\n\nArgs:\n    identifier\
        \ (str): The identifier of the task.\n    service: Task management service\
        \ dependency.\n\nReturns:\n    Response: Confirmation message of task deletion.\n\
        \nRaises:\n    TaskNotFoundError: If the task is not found."
      operationId: delete_task_task__identifier__delete_delete
      parameters:
      - name: identifier
        in: path
        required: true
        schema:
          type: string
          title: Identifier
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /tasks/{identifier}/progress:
    get:
      tags:
      - Tasks Management
      summary: Get task progress
      description: Get current progress for a task. Use this as fallback when WebSocket
        is unavailable.
      operationId: get_task_progress_tasks__identifier__progress_get
      parameters:
      - name: identifier
        in: path
        required: true
        schema:
          type: string
          title: Identifier
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskProgress'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Metadata:
      properties:
        task_type:
          type: string
          title: Task Type
        task_params:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Task Params
        language:
          anyOf:
          - type: string
          - type: 'null'
          title: Language
        file_name:
          anyOf:
          - type: string
          - type: 'null'
          title: File Name
        url:
          anyOf:
          - type: string
          - type: 'null'
          title: Url
        callback_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Callback Url
        duration:
          anyOf:
          - type: number
          - type: 'null'
          title: Duration
        audio_duration:
          anyOf:
          - type: number
          - type: 'null'
          title: Audio Duration
        start_time:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Start Time
        end_time:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: End Time
      type: object
      required:
      - task_type
      - task_params
      - language
      - file_name
      - url
      - callback_url
      - duration
      title: Metadata
      description: Metadata model for task information.
...
    Response:
      properties:
        identifier:
          type: string
          title: Identifier
        message:
          type: string
          title: Message
      type: object
      required:
      - identifier
      - message
      title: Response
      description: Response model for API responses.
    Result:
      properties:
        status:
          type: string
          title: Status
        result:
          title: Result
        metadata:
          $ref: '#/components/schemas/Metadata'
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
      type: object
      required:
      - status
      - result
      - metadata
      - error
      title: Result
      description: Model for a result with status, result data, metadata, and optional
        error.
...
    TaskListResponse:
      properties:
        tasks:
          items:
            $ref: '#/components/schemas/TaskSummaryResponse'
          type: array
          title: Tasks
          description: List of task summaries (current page slice)
        total:
          type: integer
          minimum: 0.0
          title: Total
          description: Total task count across all pages after filters apply
          default: 0
        page:
          type: integer
          minimum: 1.0
          title: Page
          description: Current page (1-indexed)
          default: 1
        page_size:
          type: integer
          maximum: 200.0
          minimum: 1.0
          title: Page Size
          description: Items per page (1..200)
          default: 50
      type: object
      required:
      - tasks
      title: TaskListResponse
      description: 'DTO for returning a paginated list of task summaries.


        Plan 15-ux pagination: ``tasks`` is the current page slice; ``total``

        is the un-paginated count after q/status filters apply (for

        "Page N of M" UI). Backwards-compatible — existing clients that only

        read ``tasks`` keep working.'
    TaskProgress:
      properties:
        identifier:
          type: string
          title: Identifier
        status:
          type: string
          title: Status
        progress_percentage:
          type: integer
          title: Progress Percentage
        progress_stage:
          anyOf:
          - type: string
          - type: 'null'
          title: Progress Stage
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
      type: object
      required:
      - identifier
      - status
      - progress_percentage
      - progress_stage
      title: TaskProgress
      description: Model for task progress information.
    TaskSummaryResponse:
      properties:
        identifier:
          type: string
          title: Identifier
          description: Unique identifier for the task
        status:
          type: string
          title: Status
          description: Current status of the task
        task_type:
          type: string
          title: Task Type
          description: Type of task
        file_name:
          anyOf:
          - type: string
          - type: 'null'
          title: File Name
          description: Name of the file
        url:
          anyOf:
          - type: string
          - type: 'null'
          title: Url
          description: URL of the file
        audio_duration:
          anyOf:
          - type: number
          - type: 'null'
          title: Audio Duration
          description: Duration of the audio
        language:
          anyOf:
          - type: string
          - type: 'null'
          title: Language
          description: Language of the audio
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
          description: Error message if task failed
        duration:
          anyOf:
          - type: number
          - type: 'null'
          title: Duration
          description: Execution duration in seconds
        start_time:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Start Time
          description: Task start time
        end_time:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: End Time
          description: Task end time
      type: object
      required:
      - identifier
      - status
      - task_type
      title: TaskSummaryResponse
      description: 'DTO for returning task summary in list operations.


        Used for GET /task/all endpoint to provide a lighter response.'
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...